types: Move SentMessageResult class to types.py.
Moved SentMessageResult class definition from actions/message_send.py
to lib/types.py for convenience. Class will also be accessed by
other modules in next commits.
