+++ zerver/lib/types.py
@@ -458,3 +458,17 @@ class DirectMessageEditRequest:
 class Invitee:
     email: str
     full_name: str = ""
+
+
+@dataclass
+class SentMessageResult:
+    """Changes to this class must be done carefully to
+    support message_response_deserializer correctly
+    deserializing values previously serialized by
+    message_response_serializer in previous versions of the server.
+    """
+
+    message_id: int
+    message_url: str
+    message_link: str
+    automatic_new_visibility_policy: int | None = None
